+++ ci/ref-from-lock.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-flakeLock="${1:?missing flake.lock}"
-input="${2:?missing input}"
+nixpkgsFrom="${1:?missing nixpkgsFrom}"
+
+flakeLock="$(echo "${nixpkgsFrom}" | cut -d# -f1)"
+input="$(echo "${nixpkgsFrom}" | cut -d# -f2)"
 
 jq -r <"${flakeLock}/flake.lock" '.nodes.root.inputs."'"${input}"'" as $name | .nodes | getpath([$name]).locked | "\(.type):\(.owner)/\(.repo)/\(.rev)"'