+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.16.0] - 2024-01-18
+
 ### Changed
 * **Breaking**: dropped compatibility for Nix versions below 2.18.1
 * **Breaking**: dropped compatibility for nixpkgs-23.05.
@@ -558,6 +560,7 @@ files parsed as nix attribute sets.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.16.0]: https://github.com/ipetkov/crane/compare/v0.15.1...v0.16.0
 [0.15.1]: https://github.com/ipetkov/crane/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/ipetkov/crane/compare/v0.14.3...v0.15.0
 [0.14.3]: https://github.com/ipetkov/crane/compare/v0.14.2...v0.14.3